[FIXED] Gradle build again
I've fixed the issue in your F-Droid metadata file (ink.trmnl.android.yml). The problem was that F-Droid was trying to run a task called assembleFdroidReleaseRelease which doesn't exist in your project. This was happening because F-Droid combines your flavor name with the word "Release" twice - once for the task name and once for the build type.