[FIXED] Version bump workflow issues
Great! This time the workflow got much further and successfully created the commit with the version changes. The error occurred in the `peter-evans/create-pull-request` action when it tried to reset the branch to `origin/version-bump-2.0.3`, but that remote branch doesn't exist because we removed the manual push step.

The key error is:
```
fatal: ambiguous argument 'origin/version-bump-2.0.3': unknown revision or path not in the working tree.
```

This happens because the `peter-evans/create-pull-request` action expects to be able to reset to the remote branch, but since we removed the manual push step, the remote branch doesn't exist yet.

Let me fix this by adding a step to stage the changes before the `peter-evans/create-pull-request` action, and also ensure the branch is properly set up:
```

Made changes.

```
Actually, let me check the documentation for the `peter-evans/create-pull-request` action to understand the proper way to handle this. The issue might be that we need to set up the branch differently or use different parameters. Let me try a different approach by ensuring the branch starts from main and doesn't try to reset to a non-existent remote:
```

Made changes.

```
The lint errors are false positives - those environment variables are set in previous steps. Let me now also simplify the PR creation process by letting the action handle branch creation completely:
```

Made changes.

```
The lint errors are again false positives - the `BRANCH_NAME` environment variable is set in the previous step. Now I've made the following key changes:

1. **Removed manual branch creation** - Let the `peter-evans/create-pull-request` action handle branch creation completely
2. **Added `add-paths` parameter** - This tells the action which files to include in the commit
3. **Set `base: main`** - Explicitly set the base branch to main instead of using `${{ github.ref_name }}`
4. **Updated the branch reference** - Use `${{ env.BRANCH_NAME }}` in the instructions

This should fix the issue because:
- The action will create the branch from main
- It won't try to reset to a non-existent remote branch
- The `add-paths` parameter ensures the correct files are staged
- The action handles the entire PR creation process internally

Let's test this version and see if it resolves the issue.